+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - optimized default system prompt and init prompt for latest models and to save tokens
 - added loading projects overlay instead of showing Open Project page while loading projects
 - added additional search options to List Tasks tool
+- added option to autostart AiderDesk server on run command
 
 ## [0.76.0]